Final changes

+++ src/main/java/com/llama/api/cart/controllers/CartController.java
@@ -1,18 +1,22 @@
 package com.llama.api.cart.controllers;
 
 import com.llama.api.cart.dto.CartItemDTO;
+import com.llama.api.cart.models.Cart;
 import com.llama.api.cart.models.CartItems;
+import com.llama.api.cart.requests.CreateCartRequest;
+import com.llama.api.cart.serializer.CartItemSerialized;
 import com.llama.api.cart.serializer.CartSerialized;
 import com.llama.api.cart.services.CartItemService;
 import com.llama.api.cart.services.CartService;
 import com.llama.api.exceptions.ResourceNotFound;
+import com.llama.api.users.models.Users;
 import com.llama.api.users.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/website/cart")
@@ -26,29 +30,41 @@ public class CartController {
     @Autowired
     UserService userService;
 
+    @GetMapping("/")
+    public ResponseEntity<CartSerialized> getCart(@RequestParam(name = "user", required = false) String user) throws ResourceNotFound {
+        CartSerialized cart;
+
+        if (user != null) {
+            Users userModel = userService.getUser(user);
+            if (userModel.getCart() != null) {
+                cart = CartSerialized.serialize(userModel.getCart());
+            } else {
+                cart = CartSerialized.serialize(cartService.createCart(user));
+            }
+        } else {
+            cart = CartSerialized.serialize(cartService.createCart());
+        }
+
+        return ResponseEntity.ok(cart);
+    }
+
     @GetMapping("/{id}/")
-    public ResponseEntity<CartSerialized> getCart(@PathVariable("id") String id) throws ResourceNotFound {
-        return ResponseEntity.ok(cartService.getCardByIDSerialized(id));
+    public ResponseEntity<CartSerialized> getCartByID(@PathVariable("id") String id) throws ResourceNotFound {
+        return ResponseEntity.ok(cartService.getCartByIDSerialized(id));
     }
 
     @PostMapping("/")
-    public ResponseEntity<CartSerialized> createCart() {
+    public ResponseEntity<CartSerialized> createCart(@Valid @RequestBody Optional<CreateCartRequest> createCartRequest) throws ResourceNotFound {
         CartSerialized cartSerialized;
 
-        try {
-            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            String username = authentication.getName();
+        if (createCartRequest.isPresent()) {
+            String userID = createCartRequest.get().getUserID();
 
             cartSerialized = CartSerialized
                     .serialize(
-                            cartService.createCart(
-                                    userService
-                                            .getUserByUsername(username)
-                                            .getId()
-                                            .toString()
-                            )
+                            cartService.createCart(userID)
                     );
-        } catch (Exception e) {
+        } else {
             cartSerialized = CartSerialized
                     .serialize(
                             cartService.createCart()
@@ -65,7 +81,6 @@ public class CartController {
                 cartItemDTO.getProductID(),
                 cartItemDTO.getQuantity()
         );
-
         return ResponseEntity.ok(
                 CartSerialized.serialize(
                         cartItem.getCart()
@@ -77,7 +92,13 @@ public class CartController {
     public ResponseEntity<CartSerialized> removeFromCart(@PathVariable("id") String id) throws ResourceNotFound {
         CartItems item = cartItemService.getItem(id);
 
-        cartItemService.deleteItem(id);
+        if (item.getQuantity() > 1) {
+            cartItemService.updateItem(item, item.getQuantity() - 1);
+        } else {
+            String cartID = item.getCart().getId().toString();
+            cartItemService.deleteItem(id);
+            cartService.updateCart(cartID);
+        }
 
         return ResponseEntity.ok(
                 CartSerialized
